+++ src/app/mod.rs
@@ -1,2 +1,3 @@
 pub mod ddns;
+pub mod playlists;
 pub mod thumbnails;
