+++ src/app/playlists/error.rs
@@ -1,0 +1,15 @@
+#[derive(thiserror::Error, Debug)]
+pub enum Error {
+	#[error("User not found")]
+	UserNotFound,
+	#[error("Playlist not found")]
+	PlaylistNotFound,
+	#[error("Unspecified")]
+	Unspecified,
+}
+
+impl From<anyhow::Error> for Error {
+	fn from(_: anyhow::Error) -> Self {
+		Error::Unspecified
+	}
+}
