+++ src/app/playlists/mod.rs
@@ -1,0 +1,7 @@
+mod error;
+mod manager;
+#[cfg(test)]
+mod test;
+
+pub use error::*;
+pub use manager::*;
